+++ packages/colonizethis_map/lib/src/tile_map_generator.dart
@@ -19,6 +19,9 @@ import 'topology_inference.dart';
 part 'tile_map_generator_types.dart';
 part 'tile_map_grid_graph.dart';
 part 'tile_map_generator_join_sea.dart';
+part 'tile_map_generator_join_sea_bridge_part.dart';
+part 'tile_map_generator_join_sea_jitter_part.dart';
+part 'tile_map_generator_join_sea_subdivide_part.dart';
 part 'tile_map_generator_terrain_assign.dart';
 part 'tile_map_generator_lakes_provinces.dart';
